Move List & Map pages to own folder 'pages'
